Move awyisser module into separate file and rename slackbot_awyiss file
